requirements.txt
# 基础Web依赖
flask==2.3.3
flask-cors==4.0.0

# 异步与API
requests==2.31.0
aiohttp==3.8.6
tenacity==8.2.3

# 数据处理与分析
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# 机器学习/深度学习
pydantic==2.4.2
torch==2.0.1
torchvision==0.15.2

# 图像处理
opencv-python==4.8.1.78
Pillow==10.0.1
mediapipe==0.10.7

# TOML配置（Python 3.11及以下）
tomllib; python_version<'3.11'

# OpenManus核心（已集成为本地包，无需单独安装）
# openmanus_core/ 目录下为本地集成代码
